backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["teteteo619 <120061560+aratetete0619@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.68.1"
uvicorn = "0.15.0"
graphene = "3.0"
starlette-graphene3 = "0.6.0"
neo4j-driver = "4.3.1"
elasticsearch = ">=7.0.0,<8.0.0"
spacy = ">=3.2.0,<3.5.0"
sudachipy = ">=0.6.2,<0.7.0"
sudachidict-core = "^20230110"
ginza = "5.1.2"
ja-ginza = "5.1.2"
python-json-logger = "2.0.2"
marshmallow = "3.13.0"
langdetect = "1.0.9"
faker = "^19.2.0"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz"}
sqlalchemy = "^2.0.19"
mysql-connector-python = "^8.1.0"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"
slowapi = "^0.1.8"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
